requirements.txt
opencv-python
tensorflow-lite-runtime
Flask
RPi.GPIO
numpy
fastapi
uvicorn
